Run prettier

+++ contracts/interfaces/yearn.sol
@@ -4,22 +4,29 @@ pragma experimental ABIEncoderV2;
 
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
 
-
 interface ICurveStrategyProxy {
     function approveStrategy(address _gauge, address _strategy) external;
-    
+
     function lock() external;
 
     function revokeStrategy(address _gauge) external;
 
-    function withdraw(address _gauge, address _token, uint _amount) external returns (uint);
+    function withdraw(
+        address _gauge,
+        address _token,
+        uint256 _amount
+    ) external returns (uint256);
 
-    function balanceOf(address _gauge) external view returns (uint);
+    function balanceOf(address _gauge) external view returns (uint256);
+
     //        return IERC20(_gauge).balanceOf(address(proxy));
 
-    function withdrawAll(address _gauge, address _token) external returns (uint);
+    function withdrawAll(address _gauge, address _token)
+        external
+        returns (uint256);
 
     function deposit(address _gauge, address _token) external;
+
     //        require(strategies[_gauge] == msg.sender, "!strategy");
     //        uint _balance = IERC20(_token).balanceOf(address(this));
     //        IERC20(_token).safeTransfer(address(proxy), _balance);
@@ -31,6 +38,7 @@ interface ICurveStrategyProxy {
     //        if (!success) assert(false);
 
     function harvest(address _gauge) external;
+
     //        require(strategies[_gauge] == msg.sender, "!strategy");
     //        uint _balance = IERC20(crv).balanceOf(address(proxy));
     //        proxy.execute(mintr, 0, abi.encodeWithSignature("mint(address)", _gauge));
@@ -41,5 +49,9 @@ interface ICurveStrategyProxy {
 }
 
 interface IKeep3rV1Oracle {
-    function current(address tokenIn, uint amountIn, address tokenOut) external view returns (uint amountOut);
+    function current(
+        address tokenIn,
+        uint256 amountIn,
+        address tokenOut
+    ) external view returns (uint256 amountOut);
 }
